# ApiDoc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Call Center",
        "description": "Daftar Api Call Center",
        "version": "1.0.0"
    },
...
    "servers": [
        {
            "url": "http://192.168.3.98:8008",
            "description": "Server Staging"
        },
        {
            "url": "http://localhost:8008",
            "description": "Server Local"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "paths": {
...
        "/calls": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Get All Calls",
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Insert Data Calls",
                "description": "",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "ref_call_id": {
                                        "type": "integer"
                                    },
                                    "direction_id": {
                                        "type": ""
                                    },
                                    "channel": {
                                        "type": "string"
                                    },
                                    "unique_id": {
                                        "type": "string"
                                    },
                                    "dnid": {
                                        "type": "string"
                                    },
                                    "extension": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "call_state_id": {
                                        "type": "integer"
                                    },
                                    "media_status_id": {
                                        "type": "integer"
                                    },
                                    "media_status_detail_id": {
                                        "type": "integer"
                                    },
                                    "media_status_detail_sub_id": {
                                        "type": "integer"
                                    },
                                    "call_date": {
                                        "type": "date"
                                    },
                                    "ivr_date": {
                                        "type": "date"
                                    },
                                    "queue_date": {
                                        "type": "date"
                                    },
                                    "leave_date": {
                                        "type": "date"
                                    },
                                    "transfer_date": {
                                        "type": "date"
                                    },
                                    "ringing_date": {
                                        "type": "date"
                                    },
                                    "bridge_date": {
                                        "type": "date"
                                    },
                                    "pickup_date": {
                                        "type": "date"
                                    },
                                    "disconected_date": {
                                        "type": "date"
                                    },
                                    "hangup_date": {
                                        "type": "date"
                                    },
                                    "dropcall_date": {
                                        "type": "date"
                                    },
                                    "speed_answer": {
                                        "type": "string"
                                    },
                                    "talk_duration": {
                                        "type": "string"
                                    },
                                    "call_duration": {
                                        "type": "string"
                                    },
                                    "acw_duration": {
                                        "type": "string"
                                    },
                                    "hold_duration": {
                                        "type": "string"
                                    },
                                    "mute_duration": {
                                        "type": "string"
                                    },
                                    "voicemail_date": {
                                        "type": "date"
                                    },
                                    "vmail_start_date": {
                                        "type": "date"
                                    },
                                    "vmail_finish_date": {
                                        "type": "date"
                                    },
                                    "host_address": {
                                        "type": "string"
                                    },
                                    "filename": {
                                        "type": "string"
                                    },
                                    "pbx_filename": {
                                        "type": "string"
                                    },
                                    "upload_date": {
                                        "type": "date"
                                    },
                                    "path": {
                                        "type": "string"
                                    },
                                    "contact_type_id": {
                                        "type": "integer"
                                    },
                                    "contact_name": {
                                        "type": "string"
                                    },
                                    "register_phone": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "user_login": {
                                        "type": "integer"
                                    },
                                    "user_not_available": {
                                        "type": "integer"
                                    },
                                    "user_ready_in": {
                                        "type": "integer"
                                    },
                                    "user_online_in": {
                                        "type": "integer"
                                    },
                                    "user_ready_out": {
                                        "type": "integer"
                                    },
                                    "user_online_out": {
                                        "type": "integer"
                                    },
                                    "user_ready_sms_in": {
                                        "type": "integer"
                                    },
                                    "user_ready_sms_out": {
                                        "type": "integer"
                                    },
                                    "user_inbound_email": {
                                        "type": "integer"
                                    },
                                    "user_outbound_email": {
                                        "type": "integer"
                                    },
                                    "user_voicemail": {
                                        "type": "integer"
                                    },
                                    "user_inbound_chat": {
                                        "type": "integer"
                                    },
                                    "user_paperwork": {
                                        "type": "integer"
                                    },
                                    "user_break": {
                                        "type": "integer"
                                    },
                                    "user_busy": {
                                        "type": "integer"
                                    },
                                    "user_ringing": {
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "type": "date"
                                    },
                                    "created_by": {
                                        "type": "integer"
                                    },
                                    "updated_at": {
                                        "type": "date"
                                    },
                                    "updated_by": {
                                        "type": "integer"
                                    },
                                    "read_at": {
                                        "type": "date"
                                    },
                                    "read_by": {
                                        "type": "integer"
                                    },
                                    "is_blocking": {
                                        "type": "integer"
                                    },
                                    "is_call_status_detail": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
        "/calls/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Get By Id Calls",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the Calls to get."
                    }
                ],
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Calls"
                ],
                "summary": "Update By Id Calls",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the Calls to get."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "ref_call_id": {
                                        "type": "integer"
                                    },
                                    "direction_id": {
                                        "type": ""
                                    },
                                    "channel": {
                                        "type": "string"
                                    },
                                    "unique_id": {
                                        "type": "string"
                                    },
                                    "dnid": {
                                        "type": "string"
                                    },
                                    "extension": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "call_state_id": {
                                        "type": "integer"
                                    },
                                    "media_status_id": {
                                        "type": "integer"
                                    },
                                    "media_status_detail_id": {
                                        "type": "integer"
                                    },
                                    "media_status_detail_sub_id": {
                                        "type": "integer"
                                    },
                                    "call_date": {
                                        "type": "date"
                                    },
                                    "ivr_date": {
                                        "type": "date"
                                    },
                                    "queue_date": {
                                        "type": "date"
                                    },
                                    "leave_date": {
                                        "type": "date"
                                    },
                                    "transfer_date": {
                                        "type": "date"
                                    },
                                    "ringing_date": {
                                        "type": "date"
                                    },
                                    "bridge_date": {
                                        "type": "date"
                                    },
                                    "pickup_date": {
                                        "type": "date"
                                    },
                                    "disconected_date": {
                                        "type": "date"
                                    },
                                    "hangup_date": {
                                        "type": "date"
                                    },
                                    "dropcall_date": {
                                        "type": "date"
                                    },
                                    "speed_answer": {
                                        "type": "string"
                                    },
                                    "talk_duration": {
                                        "type": "string"
                                    },
                                    "call_duration": {
                                        "type": "string"
                                    },
                                    "acw_duration": {
                                        "type": "string"
                                    },
                                    "hold_duration": {
                                        "type": "string"
                                    },
                                    "mute_duration": {
                                        "type": "string"
                                    },
                                    "voicemail_date": {
                                        "type": "date"
                                    },
                                    "vmail_start_date": {
                                        "type": "date"
                                    },
                                    "vmail_finish_date": {
                                        "type": "date"
                                    },
                                    "host_address": {
                                        "type": "string"
                                    },
                                    "filename": {
                                        "type": "string"
                                    },
                                    "pbx_filename": {
                                        "type": "string"
                                    },
                                    "upload_date": {
                                        "type": "date"
                                    },
                                    "path": {
                                        "type": "string"
                                    },
                                    "contact_type_id": {
                                        "type": "integer"
                                    },
                                    "contact_name": {
                                        "type": "string"
                                    },
                                    "register_phone": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "user_login": {
                                        "type": "integer"
                                    },
                                    "user_not_available": {
                                        "type": "integer"
                                    },
                                    "user_ready_in": {
                                        "type": "integer"
                                    },
                                    "user_online_in": {
                                        "type": "integer"
                                    },
                                    "user_ready_out": {
                                        "type": "integer"
                                    },
                                    "user_online_out": {
                                        "type": "integer"
                                    },
                                    "user_ready_sms_in": {
                                        "type": "integer"
                                    },
                                    "user_ready_sms_out": {
                                        "type": "integer"
                                    },
                                    "user_inbound_email": {
                                        "type": "integer"
                                    },
                                    "user_outbound_email": {
                                        "type": "integer"
                                    },
                                    "user_voicemail": {
                                        "type": "integer"
                                    },
                                    "user_inbound_chat": {
                                        "type": "integer"
                                    },
                                    "user_paperwork": {
                                        "type": "integer"
                                    },
                                    "user_break": {
                                        "type": "integer"
                                    },
                                    "user_busy": {
                                        "type": "integer"
                                    },
                                    "user_ringing": {
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "type": "date"
                                    },
                                    "created_by": {
                                        "type": "integer"
                                    },
                                    "updated_at": {
                                        "type": "date"
                                    },
                                    "updated_by": {
                                        "type": "integer"
                                    },
                                    "read_at": {
                                        "type": "date"
                                    },
                                    "read_by": {
                                        "type": "integer"
                                    },
                                    "is_blocking": {
                                        "type": "integer"
                                    },
                                    "is_call_status_detail": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
...
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ]
}